break out data points and fix data selector descriptions
