+++ src/components/municipalities/rankedList/MunicipalityDataSelector.tsx
@@ -1,106 +1,29 @@
 import React, { useEffect, useRef } from "react";
 import { BarChart3, ChevronDown } from "lucide-react";
 import { t } from "i18next";
-import { Municipality } from "@/types/municipality";
-
-interface DataPoint {
-  label: string;
-  key: keyof Municipality;
-  unit: string;
-  description?: string;
-  higherIsBetter: boolean;
-}
+import { DataPoint, dataPoints } from "./dataPoints";
 
 interface DataSelectorProps {
   selectedDataPoint: DataPoint;
   onDataPointChange: (dataPoint: DataPoint) => void;
 }
 
-export const dataPoints: DataPoint[] = [
-  {
-    label: t(
-      "municipalities.list.dataSelector.dataPoints.historicalEmissionChange.label",
-    ),
-    key: "historicalEmissionChangePercent",
-    unit: "%",
-    description: t(
-      "dataSelector.dataPoints.historicalEmissionChange.description",
-    ),
-    higherIsBetter: false,
-  },
-  {
-    label: t(
-      "municipalities.list.dataSelector.dataPoints.requiredEmissionChange.label",
-    ),
-    key: "neededEmissionChangePercent",
-    unit: "%",
-    description: t(
-      "dataSelector.dataPoints.requiredEmissionChange.description",
-    ),
-    higherIsBetter: false,
-  },
-  {
-    label: t(
-      "municipalities.list.dataSelector.dataPoints.electricCarGrowth.label",
-    ),
-    key: "electricCarChangePercent",
-    unit: "%",
-    description: t(
-      "municipalities.list.dataSelector.dataPoints.electricCarGrowth.description",
-    ),
-    higherIsBetter: true,
-  },
-  {
-    label: t(
-      "municipalities.list.dataSelector.dataPoints.totalConsumptionEmission.label",
-    ),
-    key: "totalConsumptionEmission",
-    unit: "t",
-    description: t(
-      "dataSelector.dataPoints.totalConsumptionEmission.description",
-    ),
-    higherIsBetter: false,
-  },
-  {
-    label: t(
-      "municipalities.list.dataSelector.dataPoints.evsPerChargePoint.label",
-    ),
-    key: "electricVehiclePerChargePoints",
-    unit: "",
-    description: t(
-      "municipalities.list.dataSelector.dataPoints.evsPerChargePoint.description",
-    ),
-    higherIsBetter: false,
-  },
-  {
-    label: t(
-      "municipalities.list.dataSelector.dataPoints.bicycleInfrastructure.label",
-    ),
-    key: "bicycleMetrePerCapita",
-    unit: "m",
-    description: t(
-      "municipalities.list.dataSelector.dataPoints.bicycleInfrastructure.description",
-    ),
-    higherIsBetter: true,
-  },
-];
-
-function DataSelector({
+const DataSelector = ({
   selectedDataPoint,
   onDataPointChange,
-}: DataSelectorProps) {
+}: DataSelectorProps) => {
   const [isOpen, setIsOpen] = React.useState(false);
   const dropdownRef = useRef<HTMLDivElement>(null);
 
   useEffect(() => {
-    function handleClickOutside(event: MouseEvent) {
+    const handleClickOutside = (event: MouseEvent) => {
       if (
         dropdownRef.current &&
         !dropdownRef.current.contains(event.target as Node)
       ) {
         setIsOpen(false);
       }
-    }
+    };
 
     document.addEventListener("mousedown", handleClickOutside);
     return () => document.removeEventListener("mousedown", handleClickOutside);
@@ -148,7 +71,7 @@ function DataSelector({
                   <span className="font-medium">{dataPoint.label}</span>
                   {dataPoint.description && (
                     <p className="text-xs text-white/50 mt-1">
-                      {dataPoint.description}
+                      {t(dataPoint.description)}
                     </p>
                   )}
                 </button>
@@ -159,6 +82,6 @@ function DataSelector({
       </div>
     </div>
   );
-}
+};
 
 export default DataSelector;
